IMPAKT BETA v0.1.14 RELEASE
IMPAKT BETA v0.1.14 RELEASE

* Fixes game item title overflow
* Applied standard browser > drill down > editor UX flow to Playbook
Module
* Added Playbook API functionality
* Fixes bugs where multiple dropdown lists were opening simultaneously
in the playbook editor
* Dynamic Configuration changes, V2
* functionality for popouts in playbook editor
* Intermediate checkin for popout directive including single-popout
toggling even when there are multiple popouts; click the popout toggle
to open the popout, and when clicking the toggle for other popouts, the
currently opened popout will close and the clicked popout will open.
* filters collection items with given predicate
* re-filters items when base data changes
* UI updates
* displays correct label for popouts
* Removing paper class from the Playbook editor. It causes too much
unnecessary separation of code between the Canvas to the Field.
* I’ve folded all Paper methods and properties into Canvas, and
Editor/Preview Paper into Editor/Preview Canvas accordingly.
* Playbook editor functionality updates
* refactor playbook-editor-canvas code to move canvas controller into
directive and remove the separate controller code
* refactored playbook editor data “applicator” logic into its own
module for better handling
* applicator updates
* Cleaned up some code for Actionable objects
* Updates to Field layers and general field framework improvements
* Fixes bug where Assignment/Route layers aren’t getting added to the
field
* Fixes bug where player layers aren’t getting added to the field
* Scenario preview updates
* Formation preview updates
* Play preview updates
* Added listener logic to Modifiable objects
* fixed bug where assignments weren’t moving when formation was getting
applied
* fixed bug where assignment/routes weren’t getting added to the field
layers
* fixed bug where layers were continually added to the field on each
load, rather than being cleared first
* fixed bug where input text was still populated in the applicator
popouts
* fixed bug where applicator popouts weren’t closing when clicking on
the open text input
* UI Updates
* application initialization performance improvements
* (C)R(U)D improvements for entities where modifications listener is
disabled during initialization
* deprecated playbook editor tabs service and repointed methods to
official playbook editor service methods
* UI updates for the applicator
* functionality updates for the applicator
* Fixes https://github.com/theBull/impakt/issues/268
* Fixes a syntax error in play-preview
* Adds a ‘no-personnel’ type warning, preventing user from creating
play without a base personnel
* fixes bug where canvas/field ready events were not triggering properly
* fixes bug where changing tabs doesn’t change data in the applicator
* fixes component loading bugs
* fixes bug where formation name isn’t displaying in applicator
* Handles no scenario loaded
* Handles no data in specific popouts with message
* UI lag (still needs to be fixed) where the data doesn’t load into the
applicator some times.
* Made UI updates for better screen visibility.
* Increased visibility of sidelines in preview images
* increased overall UI visibility
* Fixed a bug where the routes aren’t added to the assignment/player
layer when initially drawing, preventing the route from being cleared
when navigating to a different tab.
* Fixed a bug where applyAssignment functionality wasn’t working
properly
* Fixed a bug where opening too many tabs forces the tabs to squeeze
and hide the close button; the tabs now wrap to a new row

+++ common/models/Layer.ts
@@ -10,6 +10,7 @@ module Common.Models {
 		public zIndex: number;
 		public layers: Common.Models.LayerCollection;
 		public visible: boolean;
+		public name: string;
 
 		constructor(actionable: Common.Interfaces.IActionable, layerType: Common.Enums.LayerTypes) {
 			if (Common.Utilities.isNullOrUndefined(actionable)) {
@@ -22,6 +23,7 @@ module Common.Models {
 			this.actionable = actionable;
 			this.type = layerType;
 			this.visible = true;
+			this.name = '';
 
 			// sub layers
 			this.layers = new Common.Models.LayerCollection();
@@ -38,29 +40,57 @@ module Common.Models {
 
 			let self = this;
 
-			let predicate = function(layer: Common.Models.Layer, index: number) {
-				return self.guid == layer.guid;
-			};
-
-			return this.guid == layer.guid || this.layers.hasElementWhich(predicate);
+			return this.guid == layer.guid || 
+				this.layers.hasElementWhich(function(layer: Common.Models.Layer, index: number) {
+					return self.guid == layer.guid;
+				});
 		}
 
-		public addLayer(layer: Common.Models.Layer): void {
-			this.layers.listen(false);
-			if(this.hasLayers())
-				this.layers.add(layer);
+		public containsLayerType(type: Common.Enums.LayerTypes): boolean {
+			if(!this.hasLayers())
+				return false;
 
-			this.actionable.graphics.set.push(layer.actionable.graphics);
+			let self = this;
+			return this.type == type || 
+				this.layers.hasElementWhich(function(layer: Common.Models.Layer, index: number) {
+					return self.type == layer.type;
+				});
+		}
+
+		public addLayer(layer: Common.Models.Layer, unique?: boolean): void {
+			this.layers.listen(false);
+			if(this.hasLayers()) {
+				this.layers.add(layer);
+			}
+
+			if(this.hasGraphics())
+				this.actionable.graphics.set.push(layer.actionable.graphics);
+
 			this.layers.listen(true);
 		}
 
-		public removeLayer(layer: Common.Models.Layer): Common.Models.Layer {
-			if (this.hasGraphics())
-				layer.actionable.graphics.remove();
-
-			this.actionable.graphics.set.exclude(layer.actionable.graphics);
-
-			return this.hasLayers() ? this.layers.remove(layer.guid) : null;
+		public removeLayer(layer: Common.Models.Layer): void {
+			if (layer.guid == this.guid) {
+				this.remove();
+			} else if(this.hasLayers()) {
+				// search sub layers
+				this.layers.forEach(function(subLayer: Common.Models.Layer, index: number) {
+					// layer exists somewhere down the chain...
+					subLayer.removeLayer(layer);					
+				});
+			}
+		}
+		public removeLayerByType(type: Common.Enums.LayerTypes): void {
+			if(type == this.type) {
+				this.remove();
+			} else if(this.hasLayers()) {
+				this.layers.forEach(function(layer: Common.Models.Layer, index: number) {
+					layer.removeLayerByType(type);
+				});
+				this.layers.removeWhere(function(layer: Common.Models.Layer) {
+					return layer.type == type;
+				});
+			}
 		}
 
 		public removeAllLayers(): void {
@@ -93,7 +123,10 @@ module Common.Models {
 
 		public show(): void {
 			this.visible = true;
-			this.actionable.graphics.show();
+			
+			if(this.hasGraphics())
+				this.actionable.graphics.show();
+			
 			this.showLayers();
 		}
 
@@ -107,7 +140,10 @@ module Common.Models {
 
 		public hide(): void {
 			this.visible = false;
-			this.actionable.graphics.hide();
+			
+			if(this.hasGraphics())
+				this.actionable.graphics.hide();
+
 			this.hideLayers();
 		}
 
@@ -124,12 +160,15 @@ module Common.Models {
 		}
 
 		public removeGraphics(): void {
-			if (this.hasGraphics())
+			if(this.hasGraphics()) {
 				this.actionable.graphics.remove();
+			}
 		}
 
 		public moveByDelta(dx: number, dy: number) {
-			this.actionable.graphics.moveByDelta(dx, dy);
+			if(this.hasGraphics())
+				this.actionable.graphics.moveByDelta(dx, dy);
+
 			if(this.hasLayers()) {
 				this.layers.forEach(function(layer: Common.Models.Layer, index: number) {
 					layer.moveByDelta(dx, dy);
@@ -138,7 +177,9 @@ module Common.Models {
 		}
 
 		public drop(): void {
-			this.actionable.graphics.drop();
+			if(this.hasGraphics())
+				this.actionable.graphics.drop();
+			
 			if(this.hasLayers()) {
 				this.layers.forEach(function(layer: Common.Models.Layer, index: number) {
 					layer.drop();
@@ -173,7 +214,8 @@ module Common.Models {
 		}
 
 		public flip(): void {
-			this.actionable.graphics.flip(this.actionable.flippable);
+			if(this.hasGraphics())
+				this.actionable.graphics.flip(this.actionable.flippable);
 
 			if(this.hasLayers()) {
 				this.layers.forEach(function(layer: Common.Models.Layer, index: number) {
